+++ HISTORY.md
@@ -1,3 +1,10 @@
+v0.2.0 - 2014/04/09
+--------------------
+
+* Replaced `account.perform!` with `account.like!`, `account.subscribe_to!`
+* Added `account.add_to!` to add a video to an account’s playlist
+* Added `account.find_or_create_playlist_by` to find or create an account’s playlist
+
 v0.1.1 - 2014/04/09
 --------------------